refactor: merge movement_kernel into group.py, fix circular import
movement_group.py is deleted; movement_kernel now lives alongside kernel
in logical/group.py, referencing gemini_common.dialects.movement at
decoration time (safe because common is fully initialized before logical).

bloqade/lanes/__init__.py is converted to lazy __getattr__ exports to
break the circular import: movement/stmts.py imports
bloqade.lanes.bytecode.encoding, which previously triggered eager imports
in lanes/__init__ that transitively loaded bloqade.gemini.logical before
bloqade.gemini.common finished initializing. TYPE_CHECKING guards
preserve pyright visibility of the re-exported names.

movement/_interface.py: use list[T] annotations instead of IList[T, Len]
so that plain Python lists in @movement_kernel bodies pass pyright.
