+++ python/bloqade/gemini/common/dialects/__init__.py
@@ -1,1 +1,1 @@
-from . import qubit as qubit
+from . import movement as movement, qubit as qubit
