+++ python/bloqade/gemini/common/dialects/movement/_dialect.py
@@ -1,0 +1,3 @@
+from kirin import ir
+
+dialect = ir.Dialect("gemini.common.movement")
