+++ python/bloqade/gemini/logical/group.py
@@ -14,6 +14,7 @@ from kirin.validation import ValidationSuite
 from typing_extensions import Doc
 
 from bloqade.gemini import common as gemini_common
+from bloqade.lanes.arch.spec import ArchSpec
 
 from .dialects import operations
 
@@ -109,6 +110,122 @@ def kernel(self):
             )
             validation_result = validator.validate(mt)
             validation_result.raise_if_invalid()
+            mt.verify()
+
+    return run_pass
+
+
+@ir.dialect_group(kernel.union([gemini_common.dialects.movement]))
+def movement_kernel(self):
+    """Gemini logical kernel with user-directed atom movement.
+
+    Use instead of @kernel when your kernel calls movement.move_to(...).
+    """
+    address_analysis = address.AddressAnalysis(dialects=self)
+
+    def run_pass(
+        mt,
+        *,
+        verify: Annotated[
+            bool, Doc("run `verify` before running passes, default is `True`")
+        ] = True,
+        typeinfer: Annotated[
+            bool,
+            Doc("run type inference and apply the inferred type to IR, default `True`"),
+        ] = True,
+        fold: Annotated[bool, Doc("run folding passes")] = True,
+        aggressive: Annotated[
+            bool, Doc("run aggressive folding passes if `fold=True`")
+        ] = False,
+        inline: Annotated[bool, Doc("inline function calls, default `True`")] = True,
+        aggressive_unroll: Annotated[
+            bool,
+            Doc(
+                "Run aggressive inlining and unrolling pass on the IR, default `False`"
+            ),
+        ] = False,
+        no_raise: Annotated[bool, Doc("do not raise exception during analysis")] = True,
+        num_physical_qubits: Annotated[
+            int, Doc("number of physical qubits per logical qubit")
+        ] = 7,
+        arch_spec: Annotated[
+            ArchSpec | None, Doc("architecture spec for MoveToValidation")
+        ] = None,
+    ) -> None:
+        # stop circular import problems
+        from .rewrite.qubit_count import InsertQubitCount
+
+        if arch_spec is None:
+            from bloqade.lanes.arch.gemini.logical import get_arch_spec
+
+            arch_spec = get_arch_spec()
+
+        if inline and not aggressive_unroll:
+            InlinePass(mt.dialects, no_raise=no_raise).fixpoint(mt)
+
+        if aggressive_unroll:
+            AggressiveUnroll(mt.dialects, no_raise=no_raise).fixpoint(mt)
+        else:
+            default_pass = Default(
+                self,
+                verify=verify,
+                fold=fold,
+                aggressive=aggressive,
+                typeinfer=typeinfer,
+                no_raise=no_raise,
+            )
+
+            default_pass.fixpoint(mt)
+
+        if no_raise:
+            runner = address_analysis.run_no_raise
+        else:
+            runner = address_analysis.run
+
+        address_frame, _ = runner(mt)
+
+        rewrite.Walk(
+            rewrite.Chain(
+                WrapAddressAnalysis(address_frame.entries),
+                InsertQubitCount(num_physical_qubits),
+            )
+        ).rewrite(mt.code)
+
+        if verify:
+            # stop circular import problems
+            from ..common.validation.duplicate_address import (
+                DuplicateAddressValidation,
+            )
+            from ..common.validation.move_to import MoveToValidation
+            from .validation.clifford.analysis import (
+                GeminiLogicalValidation,
+            )
+            from .validation.measurement.analysis import (
+                GeminiTerminalMeasurementValidation,
+            )
+
+            validator = ValidationSuite(
+                [
+                    GeminiLogicalValidation,
+                    GeminiTerminalMeasurementValidation,
+                    FlatKernelNoCloningValidation,
+                    DuplicateAddressValidation,
+                ]
+            )
+            validation_result = validator.validate(mt)
+            validation_result.raise_if_invalid()
+
+            # MoveToValidation requires arch_spec at construction time and cannot
+            # be used as a bare class in ValidationSuite — run it separately.
+            _, move_to_errors = MoveToValidation(arch_spec=arch_spec).run(mt)
+            if move_to_errors:
+                from kirin.ir.exception import ValidationErrorGroup
+
+                raise ValidationErrorGroup(
+                    f"MoveToValidation failed with {len(move_to_errors)} error(s)",
+                    errors=move_to_errors,
+                )
+
             mt.verify()
 
     return run_pass
