+++ python/tests/gemini/test_movement_kernel.py
@@ -3,9 +3,9 @@
 import bloqade.squin as squin
 import pytest
 
-from bloqade.gemini.logical.movement_group import movement_kernel
+from bloqade.gemini.common.dialects.movement import move_to
+from bloqade.gemini.logical.group import movement_kernel
 from bloqade.lanes.bytecode.encoding import LocationAddress
-from bloqade.lanes.dialects.place import move_to
 
 
 def test_movement_kernel_exists():
